% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighten.R
\docType{package}
\name{lighten}
\alias{lighten}
\alias{lighten-package}
\title{lighten: Lightened choice of journal for publication}
\description{
`lighten` extracts data from medline, with the use IF and SIGAPS score of review
}
